fix: TensorRT crash by calling tensorRT with subprocess
