+++ Model/onnx_tensorrt/predictPytorch.py
@@ -1,7 +1,7 @@
 from ultralytics import YOLO
 
 # Load a model
-model = YOLO("yolov8n.pt")  # load a pretrained model (recommended for training)
+model = YOLO("../yolov8/yolov8n_custom.pt")  # load a pretrained model (recommended for training)
 model.info()
 
 # Use the model
