+++ app/main.py
@@ -10,10 +10,13 @@ from app.utils import dir_func
 from app.ffmpeg_func import video_preprocessing, combine_video_audio
 from app.subtitle_func import json2sub
 from app.audio_func import json2audio
-from Model.detector import detect
 
 from requests import get
 
+import shlex
+from subprocess import check_call, PIPE
+import json
+from pathlib import Path
 
 def get_session_id() -> str:
     ctx = get_script_run_ctx()
@@ -34,7 +37,11 @@ upload_path = f"app/uploaded/{user_session}/"
 dst_path = f"app/result/{user_session}/"
 tmp_path = f"app/tmp/{user_session}/"
 wav_path = f"app/audio/{user_session}/"
+tensorrtmodel_file = f"yolov8n_custom.trt"
 
+PRJ_ROOT_PATH = Path(__file__).parent.parent.absolute()
+MODEL_DIR = os.path.join(PRJ_ROOT_PATH, "Model")
+TENSORRT_DIR = os.path.join(MODEL_DIR, "onnx_tensorrt")
 
 def main():
     st.title("보행 시 장애물 안내 서비스")
@@ -70,13 +77,23 @@ def main():
                 placeholder.success("동영상 전처리 완료")
                 with st.spinner("객체 탐지 중..."):
                     if 1:  # Pytorch
+                        from Model.detector import detect
                         img_dir = os.path.join(tmp_path, "img_dir")
                         dir_func(img_dir, rmtree=True, mkdir=True)
                         frame_json = detect(preprocessed_file, user_session, conf_thres=slide_value)
                     else:  # TensorRT
-                        from Model.onnx_tensorrt.utils import BaseEngine
-                        pred = BaseEngine(engine_path='./Model/onnx_tensorrt/yolov8n_custom.trt')
-                        result_video_file, frame_json = pred.detect_video(file_name=preprocessed_file, user_session=user_session, conf=0.1, end2end=True)
+                        img_dir = os.path.join(tmp_path, "img_dir")
+                        dir_func(img_dir, rmtree=True, mkdir=True)
+                        tensorrt_file_path = os.path.join(TENSORRT_DIR, 'trtinference.py')
+                        preprocessing_file_path = os.path.join(tmp_path, 'preprocessed.mp4')
+                        tensorrt_model_file_path = os.path.join(TENSORRT_DIR, tensorrtmodel_file)
+                        cmd = f"python3.8 {tensorrt_file_path} -v {preprocessing_file_path} -e {tensorrt_model_file_path} -s {user_session} -c {slide_value}"
+                        check_call(shlex.split(cmd), universal_newlines=True)
+                        JSON_FILE = os.path.join(tmp_path, 'objdetection.json')
+                        with open(JSON_FILE, "rb") as json_file:
+                            frame_json_dict = json.load(json_file)
+                            frame_json = json.dumps(frame_json_dict)
+                        
                 placeholder.success("객체 탐지 완료")
                 json2sub(session_id=user_session, json_str=frame_json, fps=TARGET_FPS, save=True)
                 json2audio(dst_path=wav_path, json_str=frame_json, fps=TARGET_FPS, save=True)
